Require Google OAuth flow to access Admin
This change requires the user to complete a Google OAuth flow in order to access Admin. Completion of the flow results in setting a cookie that's valid for 6 hours and tied to the user's current IP/user-agent. The IP/user-agent check significantly increases the difficulty of using a stolen Admin cookie, and also prevents a stolen PostHog session cookie from being able to be used to access Admin.

We provide a hint to Google regarding the desired account, so the OAuth flow should be seamless and nearly indetectable. Thus, this flow should significantly harden access to Admin while not slowing down the user at all.